shuffle things around between utils and interface
