+++ pennylane/math/interface.py
@@ -19,10 +19,6 @@ from typing import Literal, get_args
 # pylint: disable=wrong-import-order
 import autoray as ar
 
-# pylint: disable=import-outside-toplevel
-from autograd.numpy.numpy_boxes import ArrayBox
-from autoray import numpy as np
-
 import pennylane as qml
 
 jpc_interfaces = {
@@ -75,33 +71,6 @@ INTERFACE_MAP = dict(zip(get_args(SupportedInterfaceUserInput), _mapping_output)
 
 SUPPORTED_INTERFACE_NAMES = list(INTERFACE_MAP)
 """list[str]: allowed interface strings"""
-
-
-def convert_like(tensor1, tensor2):
-    """Convert a tensor to the same type as another.
-
-    Args:
-        tensor1 (tensor_like): tensor to convert
-        tensor2 (tensor_like): tensor with corresponding type to convert to
-
-    Returns:
-        tensor_like: a tensor with the same shape, values, and dtype as ``tensor1`` and the
-        same type as ``tensor2``.
-
-    **Example**
-
-    >>> x = np.array([1, 2])
-    >>> y = tf.Variable([3, 4])
-    >>> convert_like(x, y)
-    <tf.Tensor: shape=(2,), dtype=int64, numpy=array([1, 2])>
-    """
-    interface = get_interface(tensor2)
-
-    if interface == "torch":
-        dev = tensor2.device
-        return np.asarray(tensor1, device=dev, like=interface)
-
-    return np.asarray(tensor1, like=interface)
 
 
 def get_interface(*values):
@@ -238,271 +207,6 @@ def get_deep_interface(value):
             return "numpy"
         itr = itr[0]
     return _get_interface_of_single_tensor(itr)
-
-
-def is_abstract(tensor, like=None):
-    """Returns True if the tensor is considered abstract.
-
-    Abstract arrays have no internal value, and are used primarily when
-    tracing Python functions, for example, in order to perform just-in-time
-    (JIT) compilation.
-
-    Abstract tensors most commonly occur within a function that has been
-    decorated using ``@tf.function`` or ``@jax.jit``.
-
-    .. note::
-
-        Currently Autograd tensors and Torch tensors will always return ``False``.
-        This is because:
-
-        - Autograd does not provide JIT compilation, and
-
-        - ``@torch.jit.script`` is not currently compatible with QNodes.
-
-    Args:
-        tensor (tensor_like): input tensor
-        like (str): The name of the interface. Will be determined automatically
-            if not provided.
-
-    Returns:
-        bool: whether the tensor is abstract or not
-
-    **Example**
-
-    Consider the following JAX function:
-
-    .. code-block:: python
-
-        import jax
-        from jax import numpy as jnp
-
-        def function(x):
-            print("Value:", x)
-            print("Abstract:", qml.math.is_abstract(x))
-            return jnp.sum(x ** 2)
-
-    When we execute it, we see that the tensor is not abstract; it has known value:
-
-    >>> x = jnp.array([0.5, 0.1])
-    >>> function(x)
-    Value: [0.5, 0.1]
-    Abstract: False
-    Array(0.26, dtype=float32)
-
-    However, if we use the ``@jax.jit`` decorator, the tensor will now be abstract:
-
-    >>> x = jnp.array([0.5, 0.1])
-    >>> jax.jit(function)(x)
-    Value: Traced<ShapedArray(float32[2])>with<DynamicJaxprTrace(level=0/1)>
-    Abstract: True
-    Array(0.26, dtype=float32)
-
-    Note that JAX uses an abstract *shaped* array, so although we won't be able to
-    include conditionals within our function that depend on the value of the tensor,
-    we *can* include conditionals that depend on the shape of the tensor.
-
-    Similarly, consider the following TensorFlow function:
-
-    .. code-block:: python
-
-        import tensorflow as tf
-
-        def function(x):
-            print("Value:", x)
-            print("Abstract:", qml.math.is_abstract(x))
-            return tf.reduce_sum(x ** 2)
-
-    >>> x = tf.Variable([0.5, 0.1])
-    >>> function(x)
-    Value: <tf.Variable 'Variable:0' shape=(2,) dtype=float32, numpy=array([0.5, 0.1], dtype=float32)>
-    Abstract: False
-    <tf.Tensor: shape=(), dtype=float32, numpy=0.26>
-
-    If we apply the ``@tf.function`` decorator, the tensor will now be abstract:
-
-    >>> tf.function(function)(x)
-    Value: <tf.Variable 'Variable:0' shape=(2,) dtype=float32>
-    Abstract: True
-    <tf.Tensor: shape=(), dtype=float32, numpy=0.26>
-    """
-    interface = like or get_interface(tensor)
-
-    if interface == "jax":
-        import jax
-        from jax.interpreters.partial_eval import DynamicJaxprTracer
-
-        if isinstance(
-            tensor,
-            (
-                jax.interpreters.ad.JVPTracer,
-                jax.interpreters.batching.BatchTracer,
-                jax.interpreters.partial_eval.JaxprTracer,
-            ),
-        ):
-            # Tracer objects will be used when computing gradients or applying transforms.
-            # If the value of the tracer is known, it will contain a ConcreteArray.
-            # Otherwise, it will be abstract.
-            return not isinstance(tensor.aval, jax.core.ConcreteArray)
-
-        return isinstance(tensor, DynamicJaxprTracer)
-
-    if interface == "tensorflow":
-        import tensorflow as tf
-        from tensorflow.python.framework.ops import EagerTensor
-
-        return not isinstance(tf.convert_to_tensor(tensor), EagerTensor)
-
-    # Autograd does not have a JIT
-
-    # QNodes do not currently support TorchScript:
-    #   NotSupportedError: Compiled functions can't take variable number of arguments or
-    #   use keyword-only arguments with defaults.
-    return False
-
-
-def import_should_record_backprop():  # pragma: no cover
-    """Return should_record_backprop or an equivalent function from TensorFlow."""
-    import tensorflow.python as tfpy
-
-    if hasattr(tfpy.eager.tape, "should_record_backprop"):
-        from tensorflow.python.eager.tape import should_record_backprop
-    elif hasattr(tfpy.eager.tape, "should_record"):
-        from tensorflow.python.eager.tape import should_record as should_record_backprop
-    elif hasattr(tfpy.eager.record, "should_record_backprop"):
-        from tensorflow.python.eager.record import should_record_backprop
-    else:
-        raise ImportError("Cannot import should_record_backprop from TensorFlow.")
-
-    return should_record_backprop
-
-
-def requires_grad(tensor, interface=None):
-    """Returns True if the tensor is considered trainable.
-
-    .. warning::
-
-        The implementation depends on the contained tensor type, and
-        may be context dependent.
-
-        For example, Torch tensors and PennyLane tensors track trainability
-        as a property of the tensor itself. TensorFlow, on the other hand,
-        only tracks trainability if being watched by a gradient tape.
-
-    Args:
-        tensor (tensor_like): input tensor
-        interface (str): The name of the interface. Will be determined automatically
-            if not provided.
-
-    Returns:
-        bool: whether the tensor is trainable or not.
-
-    **Example**
-
-    Calling this function on a PennyLane NumPy array:
-
-    >>> x = np.array([1., 5.], requires_grad=True)
-    >>> requires_grad(x)
-    True
-    >>> x.requires_grad = False
-    >>> requires_grad(x)
-    False
-
-    PyTorch has similar behaviour.
-
-    With TensorFlow, the output is dependent on whether the tensor
-    is currently being watched by a gradient tape:
-
-    >>> x = tf.Variable([0.6, 0.1])
-    >>> requires_grad(x)
-    False
-    >>> with tf.GradientTape() as tape:
-    ...     print(requires_grad(x))
-    True
-
-    While TensorFlow constants are by default not trainable, they can be
-    manually watched by the gradient tape:
-
-    >>> x = tf.constant([0.6, 0.1])
-    >>> with tf.GradientTape() as tape:
-    ...     print(requires_grad(x))
-    False
-    >>> with tf.GradientTape() as tape:
-    ...     tape.watch([x])
-    ...     print(requires_grad(x))
-    True
-    """
-    interface = interface or get_interface(tensor)
-
-    if interface == "tensorflow":
-        import tensorflow as tf
-
-        should_record_backprop = import_should_record_backprop()
-        return should_record_backprop([tf.convert_to_tensor(tensor)])
-
-    if interface == "autograd":
-        if isinstance(tensor, ArrayBox):
-            return True
-
-        return getattr(tensor, "requires_grad", False)
-
-    if interface == "torch":
-        return getattr(tensor, "requires_grad", False)
-
-    if interface in {"numpy", "scipy"}:
-        return False
-
-    if interface == "jax":
-        import jax
-
-        return isinstance(tensor, jax.core.Tracer)
-
-    raise ValueError(f"Argument {tensor} is an unknown object")
-
-
-def in_backprop(tensor, interface=None):
-    """Returns True if the tensor is considered to be in a backpropagation environment, it works for Autograd,
-    TensorFlow and Jax. It is not only checking the differentiability of the tensor like :func:`~.requires_grad`, but
-    rather checking if the gradient is actually calculated.
-
-    Args:
-        tensor (tensor_like): input tensor
-        interface (str): The name of the interface. Will be determined automatically
-            if not provided.
-
-    Returns:
-        bool: whether the tensor is in a backpropagation environment or not.
-
-    **Example**
-
-    >>> x = tf.Variable([0.6, 0.1])
-    >>> requires_grad(x)
-    False
-    >>> with tf.GradientTape() as tape:
-    ...     print(requires_grad(x))
-    True
-
-    .. seealso:: :func:`~.requires_grad`
-    """
-    interface = interface or get_interface(tensor)
-
-    if interface == "tensorflow":
-        import tensorflow as tf
-
-        should_record_backprop = import_should_record_backprop()
-        return should_record_backprop([tf.convert_to_tensor(tensor)])
-
-    if interface == "autograd":
-        return isinstance(tensor, ArrayBox)
-
-    if interface == "jax":
-        import jax
-
-        return isinstance(tensor, jax.core.Tracer)
-
-    if interface in {"numpy", "scipy"}:
-        return False
-
-    raise ValueError(f"Cannot determine if {tensor} is in backpropagation.")
 
 
 def _get_jax_interface_name(tapes):
